Adds a presenter notes view for macOS
Refactored the deck state to an ObservableObject singleton so it can be used to update multiple windows and doesn't have to be passed around everywhere.

Presenter notes for each slide can be added to the comments argument in the init method of a Slide.

Arrow keystrokes to change the slide or highlighted line work in both the presenter view and the presenter notes view